refactor: move macro arg parsing to parse mod
