+++ src/parse/mod.rs
@@ -1,2 +1,3 @@
+pub(crate) mod macros;
 pub(crate) mod parser;
 pub(crate) mod session;
